Rename activitysource.js and occurences in front-end
