+++ app/manager-frontend/src/api/activitysources.js
