+++ app/manager-frontend/src/contexts/BackendApi.jsx
@@ -1,6 +1,6 @@
 import { getAccountApi } from "api/account";
 import { getUserApi } from "api/user";
-import { getActivitySourceApi } from "api/activitysource";
+import { getActivitySourceApi } from "api/activitysources";
 import { getDataSourceApi } from "api/datasource";
 import ky from "ky";
 import { createContext, useCallback, useContext, useMemo } from "react";
