+++ app/manager-frontend/src/pages/UserHome.jsx
@@ -10,7 +10,7 @@ import {
 import { useUser } from "contexts/User";
 import { useTranslation } from "react-i18next";
 import { useSortingAndFiltering } from "helpers/hooks/useSortingAndFiltering";
-import { useActivitySourceList } from "api/activitysource";
+import { useActivitySourceList } from "api/activitysources";
 import { ActivitySourceSummary } from "components/ActivitySourceSummary";
 import { useState } from "react";
 import { useBackendApi } from "contexts/BackendApi";
